requirements.txt
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
numpy==1.26.4
opencv-python==4.9.0.80
matplotlib==3.8.4
scikit-learn==1.4.0
seaborn==0.13.2
pycocotools==2.0.7
Pillow==10.2.0
albumentations==1.4.18